patch: refactoring code, adding more resources and updating helm chart
