--- custom-errors/handler.go
@@ -1,9 +1,0 @@
-package custom_errors
-
-import "log"
-
-func LogFatalError(err error) {
-	if err != nil {
-		log.Fatalln(err.Error())
-	}
-}
