+++ hack/base.go
@@ -1,6 +1,6 @@
-package test_cases
+package hack
 
-func intPtr(number int32) *int32 {
+func IntPtr(number int32) *int32 {
 	var intPtr int32
 	intPtr = number
 	return &intPtr
