+++ resources/destinationrule.go
@@ -1,4 +1,4 @@
-package test_cases
+package resources
 
 import (
 	"context"
@@ -10,42 +10,42 @@ import (
 	"strings"
 )
 
-type VirtualServiceResource struct {
-	Client   v1beta1Typed.VirtualServiceInterface
-	Resource *v1beta1.VirtualService
+type DestinationRuleResource struct {
+	Client   v1beta1Typed.DestinationRuleInterface
+	Resource *v1beta1.DestinationRule
 	Error    error
 }
 
-func (r *VirtualServiceResource) GetObject() runtime.Object {
-	//fmt.Printf("%#v\n\n", r.Resource)
+func (r *DestinationRuleResource) GetObject() runtime.Object {
+	//fmt.Printf("%#v\n\n", r.ApiResource)
 	return r.Resource
 }
 
-func (r *VirtualServiceResource) GetError() error {
+func (r *DestinationRuleResource) GetError() error {
 	return r.Error
 }
 
-func (r *VirtualServiceResource) GetResourceName() string {
+func (r *DestinationRuleResource) GetResourceName() string {
 	return r.Resource.Name
 }
 
-func (r *VirtualServiceResource) GetResourceKind() string {
+func (r *DestinationRuleResource) GetResourceKind() string {
 	kind := strings.Split(fmt.Sprintf("%T", r.Resource), ".")
 	return kind[len(kind)-1 : len(kind)][0]
 }
 
-func (r *VirtualServiceResource) IsReady() bool {
+func (r *DestinationRuleResource) IsReady() bool {
 	if r.Resource.CreationTimestamp.IsZero() {
 		return false
 	}
 	return true
 }
 
-func (r *VirtualServiceResource) GetClient(namespace string) {
-	r.Client = istioClientset.NetworkingV1beta1().VirtualServices(namespace)
+func (r *DestinationRuleResource) GetClient(namespace string, clientset *ClientSets) {
+	r.Client = clientset.Istio.NetworkingV1beta1().DestinationRules(namespace)
 }
 
-func (r *VirtualServiceResource) Get() {
+func (r *DestinationRuleResource) Get() {
 	resource, err := r.Client.Get(context.TODO(), r.Resource.Name, metav1.GetOptions{})
 	if getHandler(r.Resource.Kind, r.Resource.Name, err) {
 		r.Resource = resource
@@ -53,12 +53,12 @@ func (r *VirtualServiceResource) Get() {
 	}
 	r.Error = err
 }
-func (r *VirtualServiceResource) Create() {
+func (r *DestinationRuleResource) Create() {
 	result, err := r.Client.Create(context.TODO(), r.Resource, metav1.CreateOptions{})
 	r.Error = err
 	r.Resource = result
 }
-func (r *VirtualServiceResource) Update() {
+func (r *DestinationRuleResource) Update() {
 }
-func (r *VirtualServiceResource) Delete() {
+func (r *DestinationRuleResource) Delete() {
 }
