+++ resources/gateway.go
@@ -1,4 +1,4 @@
-package test_cases
+package resources
 
 import (
 	"context"
@@ -10,42 +10,41 @@ import (
 	"strings"
 )
 
-type DestinationRuleResource struct {
-	Client   v1beta1Typed.DestinationRuleInterface
-	Resource *v1beta1.DestinationRule
+type GatewayResource struct {
+	Client   v1beta1Typed.GatewayInterface
+	Resource *v1beta1.Gateway
 	Error    error
 }
 
-func (r *DestinationRuleResource) GetObject() runtime.Object {
-	//fmt.Printf("%#v\n\n", r.Resource)
+func (r *GatewayResource) GetObject() runtime.Object {
+	//fmt.Printf("%#v\n\n", r.ApiResource)
 	return r.Resource
 }
 
-func (r *DestinationRuleResource) GetError() error {
+func (r *GatewayResource) GetError() error {
 	return r.Error
 }
 
-func (r *DestinationRuleResource) GetResourceName() string {
+func (r *GatewayResource) GetResourceName() string {
 	return r.Resource.Name
 }
 
-func (r *DestinationRuleResource) GetResourceKind() string {
+func (r *GatewayResource) GetResourceKind() string {
 	kind := strings.Split(fmt.Sprintf("%T", r.Resource), ".")
 	return kind[len(kind)-1 : len(kind)][0]
 }
-
-func (r *DestinationRuleResource) IsReady() bool {
+func (r *GatewayResource) IsReady() bool {
 	if r.Resource.CreationTimestamp.IsZero() {
 		return false
 	}
 	return true
 }
 
-func (r *DestinationRuleResource) GetClient(namespace string) {
-	r.Client = istioClientset.NetworkingV1beta1().DestinationRules(namespace)
+func (r *GatewayResource) GetClient(namespace string, clientset *ClientSets) {
+	r.Client = clientset.Istio.NetworkingV1beta1().Gateways(namespace)
 }
 
-func (r *DestinationRuleResource) Get() {
+func (r *GatewayResource) Get() {
 	resource, err := r.Client.Get(context.TODO(), r.Resource.Name, metav1.GetOptions{})
 	if getHandler(r.Resource.Kind, r.Resource.Name, err) {
 		r.Resource = resource
@@ -53,12 +52,12 @@ func (r *DestinationRuleResource) Get() {
 	}
 	r.Error = err
 }
-func (r *DestinationRuleResource) Create() {
+func (r *GatewayResource) Create() {
 	result, err := r.Client.Create(context.TODO(), r.Resource, metav1.CreateOptions{})
 	r.Error = err
 	r.Resource = result
 }
-func (r *DestinationRuleResource) Update() {
+func (r *GatewayResource) Update() {
 }
-func (r *DestinationRuleResource) Delete() {
+func (r *GatewayResource) Delete() {
 }
