+++ resources/secret.go
@@ -1,4 +1,4 @@
-package test_cases
+package resources
 
 import (
 	"context"
@@ -10,43 +10,42 @@ import (
 	"strings"
 )
 
-type ConfigMapResource struct {
-	Client   v1Typed.ConfigMapInterface
-	Resource *v1.ConfigMap
+type SecretResource struct {
+	Client   v1Typed.SecretInterface
+	Resource *v1.Secret
 	Error    error
 }
 
-func (r *ConfigMapResource) GetObject() runtime.Object {
-	//fmt.Printf("%#v\n\n", r.Resource)
+func (r *SecretResource) GetObject() runtime.Object {
+	//fmt.Printf("%#v\n\n", r.ApiResource)
 	return r.Resource
 }
 
-func (r *ConfigMapResource) GetError() error {
+func (r *SecretResource) GetError() error {
 	return r.Error
 }
 
-func (r *ConfigMapResource) GetResourceName() string {
+func (r *SecretResource) GetResourceName() string {
 	return r.Resource.Name
 }
 
-func (r *ConfigMapResource) GetResourceKind() string {
+func (r *SecretResource) GetResourceKind() string {
 	kind := strings.Split(fmt.Sprintf("%T", r.Resource), ".")
 	return kind[len(kind)-1 : len(kind)][0]
 }
 
-func (r *ConfigMapResource) IsReady() bool {
-	r.Get()
+func (r *SecretResource) IsReady() bool {
 	if r.Resource.CreationTimestamp.IsZero() {
 		return false
 	}
 	return true
 }
 
-func (r *ConfigMapResource) GetClient(namespace string) {
-	r.Client = clientset.CoreV1().ConfigMaps(namespace)
+func (r *SecretResource) GetClient(namespace string, clientset *ClientSets) {
+	r.Client = clientset.K8S.CoreV1().Secrets(namespace)
 }
 
-func (r *ConfigMapResource) Get() {
+func (r *SecretResource) Get() {
 	resource, err := r.Client.Get(context.TODO(), r.Resource.Name, metav1.GetOptions{})
 	if getHandler(r.Resource.Kind, r.Resource.Name, err) {
 		r.Resource = resource
@@ -54,17 +53,12 @@ func (r *ConfigMapResource) Get() {
 	}
 	r.Error = err
 }
-func (r *ConfigMapResource) Create() {
+func (r *SecretResource) Create() {
 	result, err := r.Client.Create(context.TODO(), r.Resource, metav1.CreateOptions{})
 	r.Error = err
 	r.Resource = result
 }
-func (r *ConfigMapResource) Update() {
-
+func (r *SecretResource) Update() {
 }
-func (r *ConfigMapResource) Delete() {
-}
-
-func rawConfigMap() {
-
+func (r *SecretResource) Delete() {
 }
