+++ resources/virtualservice.go
@@ -1,4 +1,4 @@
-package test_cases
+package resources
 
 import (
 	"context"
@@ -10,41 +10,42 @@ import (
 	"strings"
 )
 
-type GatewayResource struct {
-	Client   v1beta1Typed.GatewayInterface
-	Resource *v1beta1.Gateway
+type VirtualServiceResource struct {
+	Client   v1beta1Typed.VirtualServiceInterface
+	Resource *v1beta1.VirtualService
 	Error    error
 }
 
-func (r *GatewayResource) GetObject() runtime.Object {
-	//fmt.Printf("%#v\n\n", r.Resource)
+func (r *VirtualServiceResource) GetObject() runtime.Object {
+	//fmt.Printf("%#v\n\n", r.ApiResource)
 	return r.Resource
 }
 
-func (r *GatewayResource) GetError() error {
+func (r *VirtualServiceResource) GetError() error {
 	return r.Error
 }
 
-func (r *GatewayResource) GetResourceName() string {
+func (r *VirtualServiceResource) GetResourceName() string {
 	return r.Resource.Name
 }
 
-func (r *GatewayResource) GetResourceKind() string {
+func (r *VirtualServiceResource) GetResourceKind() string {
 	kind := strings.Split(fmt.Sprintf("%T", r.Resource), ".")
 	return kind[len(kind)-1 : len(kind)][0]
 }
-func (r *GatewayResource) IsReady() bool {
+
+func (r *VirtualServiceResource) IsReady() bool {
 	if r.Resource.CreationTimestamp.IsZero() {
 		return false
 	}
 	return true
 }
 
-func (r *GatewayResource) GetClient(namespace string) {
-	r.Client = istioClientset.NetworkingV1beta1().Gateways(namespace)
+func (r *VirtualServiceResource) GetClient(namespace string, clientset *ClientSets) {
+	r.Client = clientset.Istio.NetworkingV1beta1().VirtualServices(namespace)
 }
 
-func (r *GatewayResource) Get() {
+func (r *VirtualServiceResource) Get() {
 	resource, err := r.Client.Get(context.TODO(), r.Resource.Name, metav1.GetOptions{})
 	if getHandler(r.Resource.Kind, r.Resource.Name, err) {
 		r.Resource = resource
@@ -52,12 +53,12 @@ func (r *GatewayResource) Get() {
 	}
 	r.Error = err
 }
-func (r *GatewayResource) Create() {
+func (r *VirtualServiceResource) Create() {
 	result, err := r.Client.Create(context.TODO(), r.Resource, metav1.CreateOptions{})
 	r.Error = err
 	r.Resource = result
 }
-func (r *GatewayResource) Update() {
+func (r *VirtualServiceResource) Update() {
 }
-func (r *GatewayResource) Delete() {
+func (r *VirtualServiceResource) Delete() {
 }
